feat: refactor todos to module
