+++ src/modules/todos/todos.route.ts
@@ -1,0 +1,8 @@
+// Route builders with dynamic parameters
+const todosRoutes = {
+    list: "/dashboard/todos",
+    new: "/dashboard/todos/new",
+    edit: (id: string | number) => `/dashboard/todos/${id}/edit`,
+} as const;
+
+export default todosRoutes;
